Refactor medal handling: consolidate medal-related functions into MedalService and update components to use the new service methods
