+++ frontend/src/app/case-detail/case-detail.component.ts
@@ -18,18 +18,8 @@ import type {
   CaseCategory,
   SearchReportsBySteamIdGetResponse,
 } from '@zeitvertreib/types';
-import {
-  MedalIntegrityError,
-  isMedalBypassResponse,
-  isValidUrl,
-  calculateETA,
-  calculateMd5,
-  normalizeEtag,
-  normalizeHeaderValue,
-  concatUint8Arrays,
-  calculateCrc32c,
-  crc32cToBase64,
-} from '../utils/medal.utils';
+import { MedalIntegrityError, isValidUrl, calculateETA } from '../utils/medal.utils';
+import { MedalService } from '../services/medal.service';
 
 @Component({
   selector: 'app-case-detail',
@@ -139,6 +129,7 @@ export class CaseDetailComponent implements OnInit {
 
   private http = inject(HttpClient);
   private authService = inject(AuthService);
+  private medalService = inject(MedalService);
 
   constructor(
     private route: ActivatedRoute,
@@ -753,8 +744,13 @@ export class CaseDetailComponent implements OnInit {
     try {
       while (true) {
         try {
-          const medalSourceUrl = await this.resolveMedalSourceUrl(this.medalClipUrl);
-          const medalFile = await this.downloadMedalClipFromCorsProxy(medalSourceUrl);
+          const medalSourceUrl = await this.medalService.resolveMedalSourceUrl(this.medalClipUrl);
+          const medalFile = await this.medalService.downloadMedalClipFromCorsProxy(medalSourceUrl, {
+            onStatusMessage: (msg) => { this.medalStatusMessage = msg; },
+            onDownloadProgress: (progress) => { this.medalDownloadProgress = progress; },
+            onETA: (eta) => { this.medalETA = eta; },
+            onStartTime: (time) => { this.medalStartTime = time; },
+          });
 
           this.medalDownloadProgress = 100;
           this.medalUploadProgress = 0;
@@ -822,99 +818,6 @@ export class CaseDetailComponent implements OnInit {
       console.error('Medal clip upload failed:', error);
       alert(error instanceof Error ? error.message : 'Fehler beim Hochladen des Medal Clips');
     }
-  }
-
-  private async downloadMedalClipFromCorsProxy(
-    targetUrl: string,
-  ): Promise<{ file: File; extension: string; mimeType: string }> {
-    const normalizedTargetUrl = targetUrl.trim();
-    const corsProxyUrl = `${environment.medalCorsProxyUrl}${encodeURIComponent(normalizedTargetUrl)}`;
-    const response = await fetch(corsProxyUrl);
-    if (!response.ok) {
-      throw new Error('Fehler beim Abrufen des Medal Clips');
-    }
-
-    const pathName = new URL(normalizedTargetUrl).pathname;
-    const fileNameWithExt = pathName.split('/').pop() || 'clip.mp4';
-    const fileExtension = fileNameWithExt.includes('.') ? fileNameWithExt.split('.').pop()!.toLowerCase() : 'mp4';
-    const mimeType = fileExtension === 'webm' ? 'video/webm' : 'video/mp4';
-
-    this.medalStatusMessage = 'Video wird heruntergeladen...';
-    this.medalStartTime = Date.now();
-
-    const contentLength = response.headers.get('content-length');
-    const total = contentLength ? parseInt(contentLength, 10) : 0;
-    const reader = response.body?.getReader();
-    if (!reader) {
-      throw new Error('Fehler beim Lesen des Videos');
-    }
-
-    const chunks: Uint8Array[] = [];
-    let receivedLength = 0;
-
-    while (true) {
-      const readResult = await reader.read();
-      if (readResult.done) break;
-      chunks.push(readResult.value);
-      receivedLength += readResult.value.length;
-      if (total > 0) {
-        this.medalDownloadProgress = Math.round((receivedLength / total) * 100);
-        this.medalETA = calculateETA(this.medalStartTime, this.medalDownloadProgress);
-        this.medalStatusMessage = `Video wird heruntergeladen... (${this.medalDownloadProgress}%)`;
-      }
-    }
-
-    const downloadedBytes = concatUint8Arrays(chunks);
-    const etagHeader = response.headers.get('etag');
-    const checksumHeader = response.headers.get('x-amz-checksum-crc32c');
-
-    if (etagHeader) {
-      const expectedEtag = normalizeEtag(etagHeader);
-      const actualMd5 = await calculateMd5(chunks);
-      if (expectedEtag !== actualMd5.toLowerCase()) {
-        throw new MedalIntegrityError(
-          'Die Integritätsprüfung des Medal Clips ist fehlgeschlagen: ETag stimmt nicht mit dem MD5-Hash überein.',
-        );
-      }
-    }
-
-    if (checksumHeader) {
-      const expectedChecksum = normalizeHeaderValue(checksumHeader);
-      const actualChecksum = crc32cToBase64(calculateCrc32c(downloadedBytes));
-      if (expectedChecksum !== actualChecksum) {
-        throw new MedalIntegrityError(
-          'Die Integritätsprüfung des Medal Clips ist fehlgeschlagen: x-amz-checksum-crc32c stimmt nicht mit dem heruntergeladenen Inhalt überein.',
-        );
-      }
-    }
-
-    const videoFile = new File([downloadedBytes], fileNameWithExt, { type: mimeType });
-    return {
-      file: videoFile,
-      extension: fileExtension,
-      mimeType,
-    };
-  }
-
-  private async resolveMedalSourceUrl(targetUrl: string): Promise<string> {
-    const bypassUrl = `${environment.medalBypassApiUrl}${encodeURIComponent(targetUrl)}`;
-    const bypassResponse = await fetch(bypassUrl);
-    if (!bypassResponse.ok) {
-      throw new Error('Fehler beim Abrufen des Medal Clips');
-    }
-
-    const bypassData: unknown = await bypassResponse.json();
-    if (!isMedalBypassResponse(bypassData)) {
-      throw new Error('Ungültige Antwort vom Medal Bypass Service');
-    }
-    if (!bypassData.valid) {
-      throw new Error(bypassData.reasoning || 'Ungültige Medal.tv URL');
-    }
-    if (!bypassData.src) {
-      throw new Error('Keine Video-URL erhalten');
-    }
-
-    return bypassData.src;
   }
 
   /**
